+++ scripts/fdmmaterial.xsd
@@ -1,10 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <xsd:schema xmlns:xsd="http://www.w3.org/2001/XMLSchema" targetNamespace="http://www.ultimaker.com/material"
             xmlns="http://www.ultimaker.com/material" elementFormDefault="qualified">
-    <!-- This XSD document is currently unused by CI, because lxml doesn't run on that server.
-         If you plan to use it, make sure that it is still up-to-date. -->
-
-
     <xsd:element name="fdmmaterial">
         <xsd:complexType>
             <xsd:sequence>
